refactor: centralize scalar subquery nullability inference
Move row-cardinality rules into selectReturnsExactlyOneRow and drive
EXPR sublink nullability through getNonNullableColumns instead of
duplicating logic in ast-describe. Simplify getDescribedSubLink and fix
pg function overload keys using PostgreSQL argument types.

+++ packages/generate/src/generate.test.ts
@@ -3017,7 +3017,7 @@ test("scalar subquery wrapping sum() over empty rows should be nullable", async 
         {
           kind: "union",
           value: [
-            { kind: "type", value: "string", type: "numeric" },
+            { kind: "type", value: "string", type: "int8" },
             { kind: "type", value: "null", type: "null" },
           ],
         },
